Organized all the pane files to be in a single folder.
