+++ internal/components/db/connections.go
@@ -1,0 +1,28 @@
+package components
+
+// import (
+//   "fmt" 
+//   "github.com/charmbracelet/lipgloss/list"
+// )
+
+type DBConnection struct {
+  Name     string
+	Host     string
+	Port     int
+	Username string
+	Database string
+}
+
+type ConnectionList struct {
+  Conections []DBConnection
+}
+
+func NewConnectionList() *ConnectionList {
+  return &ConnectionList {
+    Conections: []DBConnection{},
+  }
+}
+
+func (cl *ConnectionList) AddConnection(conn DBConnection) {
+  cl.Conections = append(cl.Conections, conn)
+}
