+++ internal/components/ui/panes/layout.go
@@ -1,14 +1,12 @@
-package ui
+package panes
 
 import (
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/jdkingsbury/americano/pkg/components"
-	"github.com/jdkingsbury/americano/pkg/panes"
 )
 
 type PaneModel struct {
-	styles     components.PaneStyles
+	styles     PaneStyles
 	showBottom bool
 	width      int
 	height     int
@@ -27,7 +25,7 @@ func (m PaneModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case tea.WindowSizeMsg:
 		m.height = msg.Height
 		m.width = msg.Width
-		m.styles = components.CreatePaneStyles(m.width, m.height)
+		m.styles = CreatePaneStyles(m.width, m.height)
 	case tea.KeyMsg:
 		switch msg.String() {
 		case "q":
@@ -42,9 +40,9 @@ func (m PaneModel) View() string {
 	s := m.styles
 
 	// Render the panes
-	topLeftPane := panes.TopLeftPane(s.TopLeftPane)
-	mainPane := panes.MainPane(s.MainPane)
-	bottomPane := panes.BottomPane(s.BottomPane)
+	topLeftPane := TopLeftPane(s.TopLeftPane)
+	mainPane := MainPane(s.MainPane)
+	bottomPane := BottomPane(s.BottomPane)
 
 	// Arrange Panes
 	leftSide := lipgloss.JoinVertical(lipgloss.Top, topLeftPane)
