+++ internal/components/ui/panes/render.go
